Organize tests to mirror src directory
