+++ test/tui/log-viewer-utils.test.js
@@ -1,6 +1,6 @@
 import { describe, test } from 'node:test';
 import assert from 'node:assert';
-import { getEntryHeight } from '../src/tui/LogViewer.js';
+import { getEntryHeight } from '../../src/tui/LogViewer.js';
 
 const ensureVisible = (entries, terminalHeight, scrollOffset, index) => {
   if (entries.length === 0) return scrollOffset;
